menu ready

+++ client/src/utils/localStorage.js
@@ -28,17 +28,19 @@ export const updateProduct = (id, cant, table, setState) => {
 
 export const addProduct = (prod, table, setState) => {
     const carrito = JSON.parse(localStorage.getItem("basket")) || { table: table, products: [], total: 0 };
+    
+    const exists = carrito.products.filter((e) => e._id === prod._id);
 
-    carrito.products.push({
-        name: prod.name,
-        units: prod.units,
-        _id: prod._id,
-        priceByUnit: prod.price,
-    });
+    if (exists.length === 0) {
+        carrito.products.push({
+            name: prod.name,
+            units: prod.units || 1,
+            _id: prod._id,
+            priceByUnit: prod.price,
+        });
+        localStorage.setItem("basket", JSON.stringify(carrito));
+        setState && setState(carrito.products);
+    }
 
-    localStorage.setItem("basket", JSON.stringify(carrito));
-
-    setState && setState(carrito.products);
-
-    return carrito
+    return carrito;
 };
